提交监控系统代码

+++ lsn_monitor_end/db/lsn_db/StatusRecordTbl.py
@@ -8,14 +8,14 @@ class StatusRecordTbl(Base):
     __tablename__ = 'status_record_tbl'
 
     id = Column(Integer, primary_key=True)
-    record_time = Column(Integer(10), unique=False)
-    online_tree_num = Column(Integer(10), unique=False)
-    online_branch_num = Column(Integer(10), unique=False)
-    online_leaf_num = Column(Integer(10), unique=False)
-    total_online_device_num = Column(Integer(10), unique=False)
-    total_streaming_device_num = Column(Integer(10), unique=False)
-    total_push_channels = Column(Integer(10), unique=False)
-    total_connection_lines = Column(Integer(10), unique=False)
+    record_time = Column(Integer, unique=False)
+    online_tree_num = Column(Integer, unique=False)
+    online_branch_num = Column(Integer, unique=False)
+    online_leaf_num = Column(Integer, unique=False)
+    total_online_device_num = Column(Integer, unique=False)
+    total_streaming_device_num = Column(Integer, unique=False)
+    total_push_channels = Column(Integer, unique=False)
+    total_connection_lines = Column(Integer, unique=False)
 
     def __init__(self, record_time, online_tree_num, online_branch_num, online_leaf_num, total_online_device_num,
                  total_streaming_device_num, total_push_channels, total_connection_lines):
