+++ lsn_monitor_end/db/lsn_db/StreamRecordTbl.py
@@ -11,13 +11,13 @@ class StreamRecordTbl(Base):
 
     id = Column(Integer, primary_key=True)
     sender_node_id = Column(String(20), unique=False)
-    record_time = Column(Integer(11), unique=False)
+    record_time = Column(Integer, unique=False)
     source_node_id = Column(String(20), unique=False)
     object_device_node_id = Column(String(20), unique=False)
     object_guaji_node_id = Column(String(20), unique=False)
-    begin_time = Column(Integer(11), unique=False)
+    begin_time = Column(Integer, unique=False)
     end_time = Column(String(11), unique=False)
-    stream_flow = Column(Integer(8), unique=False)
+    stream_flow = Column(Integer, unique=False)
 
     def __init__(self, record_time, sender_node_id, source_node_id, object_device_node_id,
                  object_guaji_node_id, begin_time, end_time, stream_flow):
